Run-From-Zip readonly file mode checks (#2230)

+++ src/WebJobs.Script.WebHost/WebHostResolver.cs
@@ -114,7 +114,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost
                         _activeHostManager = new WebScriptHostManager(_activeScriptHostConfig, _secretManagerFactory, _eventManager, _settingsManager, settings,
                             _router, loggerProviderFactory: _loggerProviderFactory, loggerFactory: _loggerFactory);
                         //_activeReceiverManager = new WebHookReceiverManager(_activeHostManager.SecretManager);
-                        InitializeFileSystem();
+                        InitializeFileSystem(_settingsManager.FileSystemIsReadOnly);
 
                         if (_standbyHostManager != null)
                         {
@@ -150,7 +150,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost
                             _router, loggerProviderFactory: _loggerProviderFactory, loggerFactory: _loggerFactory);
                         // _standbyReceiverManager = new WebHookReceiverManager(_standbyHostManager.SecretManager);
 
-                        InitializeFileSystem();
+                        InitializeFileSystem(_settingsManager.FileSystemIsReadOnly);
                         StandbyManager.Initialize(_standbyScriptHostConfig, logger);
 
                         // start a background timer to identify when specialization happens
@@ -249,7 +249,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost
             _activeHostManager?.RunAsync(CancellationToken.None);
         }
 
-        private static void InitializeFileSystem()
+        private static void InitializeFileSystem(bool readOnlyFileSystem)
         {
             if (ScriptSettingsManager.Instance.IsAzureEnvironment)
             {
@@ -258,20 +258,26 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost
                     string home = ScriptSettingsManager.Instance.GetSetting(EnvironmentSettingNames.AzureWebsiteHomePath);
                     if (!string.IsNullOrEmpty(home))
                     {
-                        // Delete hostingstart.html if any. Azure creates that in all sites by default
-                        string siteRootPath = Path.Combine(home, "site", "wwwroot");
-                        string hostingStart = Path.Combine(siteRootPath, "hostingstart.html");
-                        if (File.Exists(hostingStart))
+                        if (!readOnlyFileSystem)
                         {
-                            File.Delete(hostingStart);
+                            // Delete hostingstart.html if any. Azure creates that in all sites by default
+                            string siteRootPath = Path.Combine(home, @"site\wwwroot");
+                            string hostingStart = Path.Combine(siteRootPath, "hostingstart.html");
+                            if (File.Exists(hostingStart))
+                            {
+                                File.Delete(hostingStart);
+                            }
                         }
 
-                        // Create the tools folder if it doesn't exist
-                        string toolsPath = Path.Combine(home, "site", "tools");
-                        Directory.CreateDirectory(toolsPath);
+                        string toolsPath = Path.Combine(home, @"site\tools");
+                        if (!readOnlyFileSystem)
+                        {
+                            // Create the tools folder if it doesn't exist
+                            Directory.CreateDirectory(toolsPath);
+                        }
 
                         var folders = new List<string>();
-                        folders.Add(Path.Combine(home, @"site", "tools"));
+                        folders.Add(Path.Combine(home, @"site\tools"));
 
                         string path = Environment.GetEnvironmentVariable("PATH");
                         string additionalPaths = string.Join(";", folders);
